Cleanup and remove old command code.
