+++ src/main/java/com/onarandombox/MultiversePortals/commands/DebugCommand.java
@@ -1,44 +1,39 @@
-/*
- * Multiverse 2 Copyright (c) the Multiverse Team 2011.
- * Multiverse 2 is licensed under the BSD License.
- * For more information please check the README.md file included
- * with this project
- */
-
 package com.onarandombox.MultiversePortals.commands;
 
-//public class DebugCommand extends PortalCommand {
-//
-//    public DebugCommand(MultiversePortals plugin) {
-//        super(plugin);
-//        this.setName("Portal Debug Mode");
-//        this.setCommandUsage("/mvp debug" + ChatColor.GOLD + " [on|off]");
-//        this.setArgRange(0, 1);
-//        this.addKey("mvp debug");
-//        this.addKey("mvpd");
-//        this.addKey("mvpdebug");
-//        this.setPermission("multiverse.portal.debug", "Instead of teleporting you to a place when you walk into a portal you will see the details about it. This command toggles.", PermissionDefault.OP);
-//    }
-//
-//    @Override
-//    public void runCommand(CommandSender sender, List<String> args) {
-//        Player p = null;
-//        if (!(sender instanceof Player)) {
-//            sender.sendMessage("This command must be run by a player");
-//            return;
-//        }
-//        p = (Player) sender;
-//
-//        if (!this.plugin.getCore().getMVWorldManager().isMVWorld(p.getWorld().getName())) {
-//            this.plugin.getCore().showNotMVWorldMessage(sender, p.getWorld().getName());
-//            return;
-//        }
-//
-//        PortalPlayerSession ps = this.plugin.getPortalSession(p);
-//        if (args.size() == 1) {
-//            ps.setDebugMode(args.get(0).equalsIgnoreCase("on"));
-//            return;
-//        }
-//        ps.setDebugMode(!ps.isDebugModeOn());
-//    }
-//}
+import com.onarandombox.MultiversePortals.MultiversePortals;
+import com.onarandombox.MultiversePortals.PortalPlayerSession;
+import com.onarandombox.acf.annotation.CommandAlias;
+import com.onarandombox.acf.annotation.CommandCompletion;
+import com.onarandombox.acf.annotation.CommandPermission;
+import com.onarandombox.acf.annotation.Description;
+import com.onarandombox.acf.annotation.Optional;
+import com.onarandombox.acf.annotation.Single;
+import com.onarandombox.acf.annotation.Subcommand;
+import com.onarandombox.acf.annotation.Syntax;
+import com.onarandombox.acf.annotation.Values;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+@CommandAlias("mvp")
+public class DebugCommand extends PortalCommand {
+
+    public DebugCommand(MultiversePortals plugin) {
+        super(plugin);
+    }
+
+    @Subcommand("debug")
+    @CommandPermission("multiverse.portal.debug")
+    @Syntax("[on|off]")
+    @CommandCompletion("@toggles")
+    @Description("Instead of teleporting you to a place when you walk into a portal you will see the details about it. This command toggles.")
+    public void onDebugCommand(@NotNull PortalPlayerSession portalSession,
+
+                               @Syntax("[on|off]")
+                               @Description("Set the state of portal debug mode.")
+                               @Nullable @Optional @Single @Values("@toggles") String mode) {
+
+        portalSession.setDebugMode((mode == null)
+                ? !portalSession.isDebugModeOn()
+                : mode.equalsIgnoreCase("on"));
+    }
+}
