+++ src/main/java/com/onarandombox/MultiversePortals/commands/ModifyCommand.java
@@ -1,176 +1,61 @@
-/*
- * Multiverse 2 Copyright (c) the Multiverse Team 2011.
- * Multiverse 2 is licensed under the BSD License.
- * For more information please check the README.md file included
- * with this project
- */
-
 package com.onarandombox.MultiversePortals.commands;
 
-/**
- * Allows modification of portal location, destination and owner. NOT name at this time.
- */
+import com.onarandombox.MultiverseCore.api.MVDestination;
+import com.onarandombox.MultiversePortals.MVPortal;
+import com.onarandombox.MultiversePortals.MultiversePortals;
+import com.onarandombox.MultiversePortals.PortalPlayerSession;
+import com.onarandombox.MultiversePortals.enums.SetProperties;
+import com.onarandombox.acf.annotation.CommandAlias;
+import com.onarandombox.acf.annotation.CommandCompletion;
+import com.onarandombox.acf.annotation.CommandPermission;
+import com.onarandombox.acf.annotation.Description;
+import com.onarandombox.acf.annotation.Flags;
+import com.onarandombox.acf.annotation.Optional;
+import com.onarandombox.acf.annotation.Subcommand;
+import com.onarandombox.acf.annotation.Syntax;
+import org.bukkit.ChatColor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
+@CommandAlias("mvp")
+public class ModifyCommand extends PortalCommand {
 
-//public class ModifyCommand extends PortalCommand {
-//
-//    public ModifyCommand(MultiversePortals plugin) {
-//        super(plugin);
-//        this.setName("Modify a Portal (Set a value)");
-//        this.setCommandUsage("/mvp modify" + ChatColor.GREEN + " {PROPERTY}" + ChatColor.GOLD + " [VALUE] -p [PORTAL]");
-//        this.setArgRange(1, 4);
-//        this.addKey("mvp modify");
-//        this.addKey("mvpmodify");
-//        this.addKey("mvpm");
-//        this.setPermission("multiverse.portal.modify", "Allows you to modify all values that can be set.", PermissionDefault.OP);
-//    }
-//
-//    @Override
-//    public void runCommand(CommandSender sender, List<String> args) {
-//        Player player = null;
-//        if ((sender instanceof Player)) {
-//            player = (Player) sender;
-//        }
-//
-//        if (!validateAction(args.get(0))) {
-//            sender.sendMessage("Sorry, you cannot " + ChatColor.AQUA + "SET" + ChatColor.WHITE + " the property " +
-//                    ChatColor.DARK_AQUA + args.get(0) + ChatColor.WHITE + ".");
-//            return;
-//        }
-//
-//        final SetProperties operator = SetProperties.valueOf(args.get(0));
-//
-//        if (!validCommand(args, operator)) {
-//            sender.sendMessage("Looks like you forgot or added an extra parameter.");
-//            sender.sendMessage("Please try again or see our Wiki for help!");
-//            return;
-//        }
-//        String portalName = extractPortalName(args);
-//        MVPortal selectedPortal = null;
-//        // If they provided -p PORTALNAME, try to retrieve it
-//        if (portalName != null) {
-//            selectedPortal = this.plugin.getPortalManager().getPortal(portalName);
-//            if (selectedPortal == null) {
-//                sender.sendMessage("Sorry, the portal " + ChatColor.RED + portalName + ChatColor.WHITE + " did not exist!");
-//                return;
-//            }
-//        } else if (player == null) {
-//            // If the player is null (so it's not a player sending the command and we don't have a portalName, we
-//            // have to return here. portalName cannot be derived if we have no player.
-//            sender.sendMessage("You NEED to enter -p PORTALNAME if you're using this from a command block or console!" +
-//                               " Not running modify command.");
-//            return;
-//        }
-//        // If they didn't provide -p, then try to use their selected portal
-//        if (selectedPortal == null) {
-//            selectedPortal = this.getUserSelectedPortal(player);
-//        }
-//
-//        if (selectedPortal == null) {
-//            sender.sendMessage("You need to select a portal using " + ChatColor.AQUA + "/mvp select {NAME}");
-//            sender.sendMessage("or append " + ChatColor.DARK_AQUA + "-p {PORTAL}" + ChatColor.WHITE + " to this command.");
-//            return;
-//        } else {
-//            portalName = selectedPortal.getName();
-//        }
-//
-//        if (portalName != null) {
-//            // Simply chop off the rest, if they have loc, that's good enough!
-//            if (operator == SetProperties.loc || operator == SetProperties.location) {
-//                this.setLocation(selectedPortal, player);
-//                return;
-//            }
-//
-//            if (operator == SetProperties.dest ||
-//                operator == SetProperties.destination) {
-//                if (args.get(1).equalsIgnoreCase("here") && player != null) {
-//                    PortalPlayerSession ps = this.plugin.getPortalSession(player);
-//                    MVPortal standingIn = ps.getUncachedStandingInPortal();
-//                    Location l = player.getLocation();
-//                    if (standingIn != null) {
-//                        String cardinal = LocationManipulation.getDirection(l);
-//                        args.set(1, "p:" + standingIn.getName() + ":" + cardinal);
-//                    } else {
-//                        args.set(1, "e:" + l.getWorld().getName() + ":" + l.getX() + "," + l.getY() + "," + l.getZ() + ":" + l.getPitch() + ":" + l.getYaw());
-//                    }
-//                } else if (args.get(1).matches("(i?)cannon-[\\d]+(\\.[\\d]+)?") && player != null) {
-//                    // We found a Cannon Destination!
-//                    Location l = player.getLocation();
-//                    try {
-//                        Double speed = Double.parseDouble(args.get(1).split("-")[1]);
-//                        args.set(1, "ca:" + l.getWorld().getName() + ":" + l.getX() + "," + l.getY() + "," + l.getZ() + ":" + l.getPitch() + ":" + l.getYaw() + ":" + speed);
-//                    } catch (NumberFormatException e) {
-//                    }
-//
-//                }
-//            }
-//
-//            if (this.setProperty(selectedPortal, args.get(0), args.get(1))) {
-//                sender.sendMessage("Property " + args.get(0) + " of Portal " + ChatColor.YELLOW + selectedPortal.getName() + ChatColor.GREEN + " was set to " + ChatColor.AQUA + args.get(1));
-//            } else {
-//                sender.sendMessage("Property " + args.get(0) + " of Portal " + ChatColor.YELLOW + selectedPortal.getName() + ChatColor.RED + " was NOT set to " + ChatColor.AQUA + args.get(1));
-//                if (args.get(0).equalsIgnoreCase("dest") || args.get(0).equalsIgnoreCase("destination")) {
-//                    sender.sendMessage("Multiverse could not find the destination: " + ChatColor.GOLD + args.get(1));
-//                }
-//            }
-//        }
-//    }
-//
-//    private boolean setProperty(MVPortal selectedPortal, String property, String value) {
-//        return selectedPortal.setProperty(property, value);
-//    }
-//
-//    private void setLocation(MVPortal selectedPortal, Player player) {
-//        PortalPlayerSession ps = this.plugin.getPortalSession(player);
-//        MultiverseRegion r = ps.getSelectedRegion();
-//        if (r != null) {
-//            MultiverseWorld world = this.plugin.getCore().getMVWorldManager().getMVWorld(player.getWorld().getName());
-//            PortalLocation location = new PortalLocation(r.getMinimumPoint(), r.getMaximumPoint(), world);
-//            selectedPortal.setPortalLocation(location);
-//            player.sendMessage("Portal location has been set to your " + ChatColor.GREEN + "selection" + ChatColor.WHITE + "!");
-//        }
-//    }
-//
-//    private MVPortal getUserSelectedPortal(Player player) {
-//        PortalPlayerSession ps = this.plugin.getPortalSession(player);
-//        return ps.getSelectedPortal();
-//    }
-//
-//    private boolean validCommand(List<String> args, SetProperties property) {
-//        // This means that they did not specify the -p or forgot the [PORTAL]
-//
-//        if (property == SetProperties.loc && args.size() % 2 == 0) {
-//            return false;
-//        } else if (property != SetProperties.loc && args.size() % 2 != 0) {
-//            return false;
-//        }
-//        return true;
-//    }
-//
-//    private String extractPortalName(List<String> args) {
-//        if (!args.contains("-p")) {
-//            return null;
-//        }
-//
-//        int index = args.indexOf("-p");
-//
-//        // Now we remove the -p
-//        args.remove(index);
-//
-//        // Now we remove and return the portalname if present
-//        if (index < args.size()) {
-//            return args.remove(index);
-//        }
-//
-//        return null;
-//    }
-//
-//    protected static boolean validateAction(String property) {
-//        try {
-//            SetProperties.valueOf(property);
-//            return true;
-//        } catch (IllegalArgumentException e) {
-//            return false;
-//        }
-//    }
-//}
+    public ModifyCommand(MultiversePortals plugin) {
+        super(plugin);
+    }
+
+    @Subcommand("modify")
+    @CommandPermission("multiverse.portal.modify")
+    @Syntax("<property> <value> [portal]")
+    @CommandCompletion("@portalProperties @empty @MVPortals")
+    @Description("Modify portal properties.")
+    public void onModifyCommand(@NotNull CommandSender sender,
+                                @Nullable @Optional Player player,
+                                @Nullable @Optional PortalPlayerSession portalSession,
+                                @NotNull SetProperties property,
+                                @NotNull @Flags("type=property value") String value,
+                                @NotNull @Flags("defaultself") MVPortal portal) {
+
+        if (property == SetProperties.destination) {
+            if (value.equalsIgnoreCase("here") && player != null) {
+                portal.setHereDestination(player, portalSession);
+                return;
+            }
+
+            MVDestination destination = this.plugin.getCore()
+                    .getDestFactory()
+                    .getPlayerAwareDestination(value, player);
+
+            sender.sendMessage((portal.setDestination(destination))
+                    ? "Destination of new portal is " + ChatColor.AQUA + destination.toString() + ChatColor.WHITE + "."
+                    : ChatColor.RED + "There was an error setting Destination " + ChatColor.AQUA + destination.toString()
+                    + ChatColor.RED + ". Is it formatted correctly?");
+        }
+
+        sender.sendMessage((portal.setProperty(property.toString(), value))
+                ? "Property " + property.toString() + " of Portal " + ChatColor.YELLOW + portal.getName() + ChatColor.GREEN + " was set to " + ChatColor.AQUA + value
+                : "Error! Property " + property.toString() + " of Portal " + ChatColor.YELLOW + portal.getName() + ChatColor.RED + " could not be set to " + ChatColor.AQUA + value);
+    }
+}
