+++ src/main/java/com/onarandombox/MultiversePortals/utils/CommandTools.java
@@ -8,21 +8,20 @@
 package com.onarandombox.MultiversePortals.utils;
 
 import com.onarandombox.MultiverseCore.commandTools.MVCommandManager;
-import com.onarandombox.MultiverseCore.utils.webpaste.PasteServiceType;
 import com.onarandombox.MultiversePortals.MVPortal;
 import com.onarandombox.MultiversePortals.MultiversePortals;
 import com.onarandombox.MultiversePortals.PortalPlayerSession;
-import com.onarandombox.MultiversePortals.commands_acf.ConfigCommand;
-import com.onarandombox.MultiversePortals.commands_acf.CreateCommand;
-import com.onarandombox.MultiversePortals.commands_acf.DebugCommand;
-import com.onarandombox.MultiversePortals.commands_acf.InfoCommand;
-import com.onarandombox.MultiversePortals.commands_acf.ListCommand;
-import com.onarandombox.MultiversePortals.commands_acf.ModifyCommand;
-import com.onarandombox.MultiversePortals.commands_acf.RemoveCommand;
-import com.onarandombox.MultiversePortals.commands_acf.RootCommand;
-import com.onarandombox.MultiversePortals.commands_acf.SelectCommand;
-import com.onarandombox.MultiversePortals.commands_acf.UsageCommand;
-import com.onarandombox.MultiversePortals.commands_acf.WandCommand;
+import com.onarandombox.MultiversePortals.commands.ConfigCommand;
+import com.onarandombox.MultiversePortals.commands.CreateCommand;
+import com.onarandombox.MultiversePortals.commands.DebugCommand;
+import com.onarandombox.MultiversePortals.commands.InfoCommand;
+import com.onarandombox.MultiversePortals.commands.ListCommand;
+import com.onarandombox.MultiversePortals.commands.ModifyCommand;
+import com.onarandombox.MultiversePortals.commands.RemoveCommand;
+import com.onarandombox.MultiversePortals.commands.RootCommand;
+import com.onarandombox.MultiversePortals.commands.SelectCommand;
+import com.onarandombox.MultiversePortals.commands.UsageCommand;
+import com.onarandombox.MultiversePortals.commands.WandCommand;
 import com.onarandombox.MultiversePortals.enums.PortalConfigProperty;
 import com.onarandombox.MultiversePortals.enums.SetProperties;
 import com.onarandombox.acf.BukkitCommandCompletionContext;
